chore: replace hacky useLayoutEffect animation work-around for filter
