--- app/(map)/components/sidebar/menu/filter/category.module.css
@@ -1,76 +1,0 @@
-.category,
-.entries {
-    display: flex;
-    flex-direction: column;
-}
-
-.entries {
-    display: grid;
-    grid-template-columns: repeat(2, minmax(0, 1fr));
-    gap: var(--gap-3xs) var(--gap-xl);
-    margin-top: var(--gap-3xs);
-    overflow: hidden;
-    align-self: stretch;
-    transition:
-        max-height 0.4s ease,
-        margin-top 0.4s ease;
-}
-
-.entries.collapsing {
-    margin-top: 0;
-}
-
-.header {
-    display: flex;
-    align-items: center;
-    align-self: stretch;
-    gap: var(--gap-6xs);
-    font-size: var(--font-size-title);
-    font-weight: 600;
-}
-
-.button,
-.caret {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0;
-    margin: 0;
-    font: inherit;
-    color: inherit;
-    cursor: pointer;
-    text-align: left;
-    outline: none;
-    background: none;
-    border: none;
-}
-
-.caret {
-    color: var(--text-dark);
-    transition:
-        color 0.2s ease,
-        transform 0.4s ease;
-}
-
-.caret.inactive {
-    color: var(--text-disabled-dark);
-}
-
-.caret.collapsed {
-    transform: scale(1, -1);
-}
-
-.button.active {
-    color: inherit;
-    text-decoration: none;
-}
-
-.button.inactive {
-    color: var(--text-disabled-dark);
-    text-decoration: line-through;
-}
-
-.button[data-bookmark-state='partial'] svg path:first-of-type {
-    fill: currentColor;
-    opacity: 0.3;
-}
